requirements.txt
python-pptx
gtts
moviepy
pydub
pdf2image
gradio_client
natsort